requirements.txt
nature-remo==0.3.3
python-dotenv==0.15.0
influxdb==5.3.1
PyYAML==5.4.1
pandas==1.2.2
numpy==1.20.1
